+++ tests/Acme81/Attribute/ParameterA.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Acme81\Attribute;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_PARAMETER)]
+class ParameterA
+{
+    public function __construct(
+        public string $label = ''
+    ) {
+    }
+}