+++ tests/Acme81/Attribute/ParameterB.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Acme81\Attribute;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_PARAMETER)]
+class ParameterB
+{
+    public function __construct(
+        public string $label = '',
+        public string $moreData = ''
+    ) {
+    }
+}