+++ tests/CollectionTest.php
@@ -18,10 +18,13 @@ use Acme\PSR4\ActiveRecord\Article;
 use Acme\PSR4\DeleteMenu;
 use Acme\PSR4\Presentation\ArticleController;
 use Closure;
+use Acme81\Attribute\ParameterA;
+use Acme81\Attribute\ParameterB;
 use olvlvl\ComposerAttributeCollector\Attributes;
 use olvlvl\ComposerAttributeCollector\Collection;
 use olvlvl\ComposerAttributeCollector\TargetClass;
 use olvlvl\ComposerAttributeCollector\TargetMethod;
+use olvlvl\ComposerAttributeCollector\TargetParameter;
 use olvlvl\ComposerAttributeCollector\TargetProperty;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
@@ -52,6 +55,8 @@ final class CollectionTest extends TestCase
                 Serial::class => [
                     [ [ 'Primary' => true ], Article::class, 'id' ],
                 ]
+            ],
+            targetParameters: [
             ]
         );
 
@@ -104,6 +109,8 @@ final class CollectionTest extends TestCase
             targetMethods: [
             ],
             targetProperties: [
+            ],
+            targetParameters: [
             ]
         );
 
@@ -134,6 +141,8 @@ final class CollectionTest extends TestCase
                 ],
             ],
             targetProperties: [
+            ],
+            targetParameters: [
             ]
         );
 
@@ -143,6 +152,36 @@ final class CollectionTest extends TestCase
             new TargetMethod(new Route('/recent'), ArticleController::class, 'recent'),
             new TargetMethod(new Get(), ArticleController::class, 'show'),
             new TargetMethod(new Post(), ArticleController::class, 'create'),
+        ], $actual);
+    }
+
+    public function testFilterTargetParameters(): void
+    {
+        $collection = new Collection(
+            targetClasses: [
+            ],
+            targetMethods: [
+            ],
+            targetProperties: [
+            ],
+            targetParameters: [
+                ParameterA::class => [
+                    [ [ 'a' ], ArticleController::class, 'myMethod', 'myParamA', ],
+                    [ [ 'a2' ], ArticleController::class, 'myMethod', 'myParamA2' ],
+                    [ [ 'a3' ], ArticleController::class, 'myFoo', 'fooParam' ],
+                ],
+                ParameterB::class => [
+                    [ [ 'b', 'more data'], ArticleController::class, 'myMethod', 'myParamB' ],
+                ],
+            ]
+        );
+
+        $actual = $collection->filterTargetParameters(fn($a) => is_a($a, ParameterA::class, true));
+
+        $this->assertEquals([
+            new TargetParameter(new ParameterA('a'), ArticleController::class, 'myMethod', 'myParamA'),
+            new TargetParameter(new ParameterA('a2'), ArticleController::class, 'myMethod', 'myParamA2'),
+            new TargetParameter(new ParameterA('a3'), ArticleController::class, 'myFoo', 'fooParam'),
         ], $actual);
     }
 
@@ -175,6 +214,8 @@ final class CollectionTest extends TestCase
                 Text::class => [
                     [ [ ], Article::class, 'body' ],
                 ]
+            ],
+            targetParameters: [
             ]
         );
 
@@ -220,6 +261,8 @@ final class CollectionTest extends TestCase
                 Text::class => [
                     [ [ ], Article::class, 'body' ],
                 ]
+            ],
+            targetParameters: [
             ]
         );