+++ tests/PluginTest.php
@@ -1,12 +1,5 @@
 <?php
 
-/*
- * (c) Olivier Laviale <olivier.laviale@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace tests\olvlvl\ComposerAttributeCollector;
 
 use Acme\Attribute\ActiveRecord\Boolean;
@@ -22,6 +15,7 @@ use Acme\Attribute\Handler;
 use Acme\Attribute\Permission;
 use Acme\Attribute\Resource;
 use Acme\Attribute\Route;
+use Acme\Attribute\Routing\UrlGetter;
 use Acme\Attribute\Subscribe;
 use Acme\PSR4\Presentation\ArticleController;
 use Acme81\Attribute\ParameterA;
@@ -48,9 +42,6 @@ final class PluginTest extends TestCase
 {
     private static bool $initialized = false;
 
-    /**
-     * @throws ReflectionException
-     */
     protected function setUp(): void
     {
         parent::setUp();
@@ -165,37 +156,43 @@ final class PluginTest extends TestCase
     {
         return [
 
+//            [
+//                Route::class,
+//                [
+//                    [
+//                        new Route("/articles/method/", 'GET', 'articles:method'),
+//                        'Acme\PSR4\Presentation\ArticleController::aMethod'
+//                    ],
+//                    [
+//                        new Route("/articles", 'GET', 'articles:list'),
+//                        'Acme\PSR4\Presentation\ArticleController::list'
+//                    ],
+//                    [
+//                        new Route("/articles/{id}", 'GET', 'articles:show'),
+//                        'Acme\PSR4\Presentation\ArticleController::show'
+//                    ],
+//                ]
+//            ],
+//            [
+//                Get::class,
+//                [
+//                    [ new Get(), 'Acme\Presentation\FileController::list' ],
+//                    [ new Get('/{id}'), 'Acme\Presentation\FileController::show' ],
+//                    [ new Get(), 'Acme\Presentation\ImageController::list' ],
+//                    [ new Get('/{id}'), 'Acme\Presentation\ImageController::show' ],
+//                ]
+//            ],
+//            [
+//                Subscribe::class,
+//                [
+//                    [ new Subscribe(), 'Acme\PSR4\SubscriberA::onEventA' ],
+//                    [ new Subscribe(), 'Acme\PSR4\SubscriberB::onEventA' ],
+//                ]
+//            ],
             [
-                Route::class,
+                UrlGetter::class,
                 [
-                    [
-                        new Route("/articles/method/", 'GET', 'articles:method'),
-                        'Acme\PSR4\Presentation\ArticleController::aMethod'
-                    ],
-                    [
-                        new Route("/articles", 'GET', 'articles:list'),
-                        'Acme\PSR4\Presentation\ArticleController::list'
-                    ],
-                    [
-                        new Route("/articles/{id}", 'GET', 'articles:show'),
-                        'Acme\PSR4\Presentation\ArticleController::show'
-                    ],
-                ]
-            ],
-            [
-                Get::class,
-                [
-                    [ new Get(), 'Acme\Presentation\FileController::list' ],
-                    [ new Get('/{id}'), 'Acme\Presentation\FileController::show' ],
-                    [ new Get(), 'Acme\Presentation\ImageController::list' ],
-                    [ new Get('/{id}'), 'Acme\Presentation\ImageController::show' ],
-                ]
-            ],
-            [
-                Subscribe::class,
-                [
-                    [ new Subscribe(), 'Acme\PSR4\SubscriberA::onEventA' ],
-                    [ new Subscribe(), 'Acme\PSR4\SubscriberB::onEventA' ],
+                    [ new UrlGetter(), 'Acme\PSR4\InheritedAttributeSample::get_url' ],
                 ]
             ],